+++ frontend/src/route.rs
@@ -1,0 +1,6 @@
+use yew::{html::Scope, prelude::*};
+use yew_router::prelude::*;
+
+pub mod login;
+pub mod main;
+pub mod settings;
